requirements.txt
advent-of-code-data==2.1.0
aocd-example-parser==2025.12.1
astroid==4.0.2
beautifulsoup4==4.14.3
cfgv==3.5.0
dill==0.4.0
distlib==0.4.0
filelock==3.20.0
identify==2.6.15
isort==7.0.0
mccabe==0.7.0
nodeenv==1.9.1
Pebble==5.1.3
platformdirs==4.5.0
pre_commit==4.5.0
pylint==4.0.4
PyYAML==6.0.3
soupsieve==2.8
tomlkit==0.13.3
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.4
